requirements.txt
gensim==4.0.1
langdetect==1.0.9
SpeechRecognition==3.8.1
Django==3.2.5
pandas==1.2.5
nltk==3.6.2
numpy==1.20.2
djangorestframework==3.12.4
PyAudio==0.2.11
